requirements.txt
fastapi
uvicorn
numpy
scikit-learn
numba
mmh3
sentence-transformers
faiss-cpu
pandas
redis
aiofiles
pydantic-settings
pytest